Browser retry ladder: fresh browser on crashes, one calm retry after a block, then API
- Maps: crash/timeout -> job retried with a new browser up to
  browser_retries (default 2), then the phrase continues on the Places
  API; CAPTCHA/block -> source quiet for block_retry_minutes (default 30),
  one more normal attempt, then Places API (fallback) or pause
- Browser search: crash -> new browser up to browser_retries, then
  Serper/Brave; block -> Serper/Brave immediately (inside a larger step)
- RetryLater postpones a job without counting a failure
- No new-identity / other-browser retries on CAPTCHAs, no solving
- 3 new tests (70 total)

+++ app/services/browser.py
@@ -139,7 +139,7 @@ def blocked_until(db, source: str) -> datetime | None:
     return until if until > datetime.now(timezone.utc) else None
 
 
-def set_blocked(db, source: str, hours: int) -> datetime:
+def set_blocked(db, source: str, hours: float) -> datetime:
     from app.models import Setting
     from app.security import encrypt
